+++ src/FSharpLint.Core/Framework/AbstractSyntaxArray.fs
@@ -129,16 +129,26 @@ module AbstractSyntaxArray =
         | InterfaceImplementation(_)
         | TypeRepresentation(_)
         | File(_)
-        | AstNode.ComponentInfo(_) -> SyntaxNode.Other
-        | AstNode.EnumCase(_) -> SyntaxNode.EnumCase
-        | AstNode.UnionCase(_) -> SyntaxNode.UnionCase
-
+        | LambdaArg(_)
+        | LambdaBody(_) 
+        | Else(_) 
+        | ComponentInfo(_) -> SyntaxNode.Other
+        | EnumCase(_) -> SyntaxNode.EnumCase
+        | UnionCase(_) -> SyntaxNode.UnionCase
+        
     [<Struct; NoEquality; NoComparison; DebuggerDisplay("{DebuggerDisplay,nq}")>]
-    type Node(hashcode: int, actual: AstNode) =
+    type TempNode(hashcode: int, actual: AstNode) =
         member __.Hashcode = hashcode
         member __.Actual = actual
-
+        
         member private __.DebuggerDisplay = "AstNode: " + string actual
+
+    [<NoEquality; NoComparison>]
+    type Node =
+        { Hashcode:int
+          Actual:AstNode
+          ParentIndex:int
+          NumberOfChildren:int }
 
     [<Struct>]
     type private PossibleSkip(skipPosition: int, depth: int) =
@@ -194,17 +204,9 @@ module AbstractSyntaxArray =
         | _ -> 0
 
     [<Struct; NoEquality; NoComparison>]
-    type private StackedNode(node: Ast.Node, depth: int) =
+    type private StackedNode(node: AstNode, depth: int) =
         member __.Node = node
         member __.Depth = depth
-
-    [<Struct; DebuggerDisplay("{DebuggerDisplay,nq}")>]
-    type Skip(numberOfChildren: int, parentIndex: int) =
-        member __.NumberOfChildren = numberOfChildren
-        member __.ParentIndex = parentIndex
-
-        member private __.DebuggerDisplay =
-            "Skip: NumberOfChildren=" + string numberOfChildren + ", ParentIndex=" + string parentIndex
 
     /// Keep index of position so skip array can be created in the correct order.
     [<Struct>]
@@ -237,61 +239,65 @@ module AbstractSyntaxArray =
                 let parentIndex = if possibleSkips.Count > 0 then possibleSkips.Peek().SkipPosition else 0
                 skips.Add(TempSkip(numberOfChildren, parentIndex, nodePosition))
 
-        left.Push (StackedNode(Ast.Node(ExtraSyntaxInfo.None, astRoot), 0))
+        left.Push (StackedNode(astRoot, 0))
 
         while left.Count > 0 do
             let stackedNode = left.Pop()
             let node = stackedNode.Node
-            let astNode = node.AstNode
             let depth = stackedNode.Depth
 
             tryAddPossibleSkips depth
+            
+            // Strip out "extra info".
+            let node =
+                let extractExtraInfo actual extraInfoNode =
+                    possibleSkips.Push (PossibleSkip(nodes.Count, depth))
+                    nodes.Add (TempNode(Utilities.hash2 extraInfoNode 0, actual))
+                    actual
 
-            traverseNode astNode (fun node -> left.Push (StackedNode(node, depth + 1)))
+                match node with
+                | LambdaArg(arg) -> extractExtraInfo (SimplePatterns(arg)) SyntaxNode.LambdaArg
+                | LambdaBody(body) -> extractExtraInfo (Expression(body)) SyntaxNode.LambdaBody
+                | Else(body) -> extractExtraInfo (Expression(body)) SyntaxNode.Else
+                | _ -> node
 
-            if node.ExtraSyntaxInfo <> ExtraSyntaxInfo.None then
-                possibleSkips.Push (PossibleSkip(nodes.Count, depth))
+            traverseNode node (fun node -> left.Push (StackedNode(node, depth + 1)))
 
-                let syntaxNode =
-                    match node.ExtraSyntaxInfo with
-                    | ExtraSyntaxInfo.LambdaArg -> SyntaxNode.LambdaArg
-                    | ExtraSyntaxInfo.LambdaBody -> SyntaxNode.LambdaBody
-                    | ExtraSyntaxInfo.Else -> SyntaxNode.Else
-                    | _ -> failwith ("Unknown extra syntax info: " + string node.ExtraSyntaxInfo)
-
-                nodes.Add (Node(Utilities.hash2 syntaxNode 0, astNode))
-
-            match astNodeToSyntaxNode astNode with
+            match astNodeToSyntaxNode node with
             | SyntaxNode.Other -> ()
             | syntaxNode ->
                 possibleSkips.Push (PossibleSkip(nodes.Count, depth))
-
-                nodes.Add (Node(Utilities.hash2 syntaxNode (getHashCode astNode), astNode))
+                nodes.Add (TempNode(Utilities.hash2 syntaxNode (getHashCode node), node))
 
         tryAddPossibleSkips 0
 
-        let skipArray = Array.zeroCreate skips.Count
+        let result = Array.zeroCreate nodes.Count
 
         let mutable i = 0
-        while i < skips.Count do
+        while i < nodes.Count do
             let skip = skips.[i]
-            skipArray.[skip.Index] <- Skip(skip.NumberOfChildren, skip.ParentIndex)
+            let node = nodes.[skip.Index]
+
+            result.[skip.Index] <- 
+                { Hashcode = node.Hashcode
+                  Actual = node.Actual
+                  NumberOfChildren = skip.NumberOfChildren
+                  ParentIndex = skip.ParentIndex }
 
             i <- i + 1
 
-        (nodes.ToArray(), skipArray)
+        result
 
-    let getBreadcrumbs maxBreadcrumbs (syntaxArray:Node []) (skipArray:Skip []) i =
+    let getBreadcrumbs maxBreadcrumbs (syntaxArray:Node []) i =
         let rec getBreadcrumbs breadcrumbs i =
             if i = 0 then
-                let node = syntaxArray.[i].Actual
-                node::breadcrumbs
-            else if i < skipArray.Length && (List.length breadcrumbs) < maxBreadcrumbs then
-                let node = syntaxArray.[i].Actual
-                let parenti = skipArray.[i].ParentIndex
-                getBreadcrumbs (node::breadcrumbs) parenti
+                let node = syntaxArray.[i]
+                node.Actual::breadcrumbs
+            else if i < syntaxArray.Length && (List.length breadcrumbs) < maxBreadcrumbs then
+                let node = syntaxArray.[i]
+                getBreadcrumbs (node.Actual::breadcrumbs) node.ParentIndex
             else
                 breadcrumbs
 
         if i = 0 then []
-        else getBreadcrumbs [] (skipArray.[i].ParentIndex) |> List.rev
+        else getBreadcrumbs [] (syntaxArray.[i].ParentIndex) |> List.rev