+++ src/FSharpLint.Core/Rules/Conventions/Binding/BindingHelper.fs
@@ -4,9 +4,9 @@ open FSharpLint.Framework
 open FSharp.Compiler.SyntaxTree
 open FSharpLint.Framework.Ast
 
-let isLetBinding i (syntaxArray:AbstractSyntaxArray.Node []) (skipArray:AbstractSyntaxArray.Skip []) =
+let isLetBinding i (syntaxArray:AbstractSyntaxArray.Node []) =
     if i > 0 then
-        match syntaxArray.[skipArray.[i].ParentIndex].Actual with
+        match syntaxArray.[syntaxArray.[i].ParentIndex].Actual with
         | AstNode.ModuleDeclaration(SynModuleDecl.Let(_))
         | AstNode.Expression(SynExpr.LetOrUse(_, false, _, _, _)) -> true
         | _ -> false