+++ src/FSharpLint.Core/Rules/Conventions/Binding/UselessBinding.fs
@@ -15,22 +15,21 @@ let private checkForUselessBinding (checkInfo:FSharpCheckFileResults option) pat
             | SynPat.Paren(pattern, _) -> findBindingIdentifier pattern
             | SynPat.Named(_, ident, _, _, _) -> Some(ident)
             | _ -> None
+        
+        let isNotMutable (symbol:FSharpSymbolUse) =
+            match symbol.Symbol with
+            | :? FSharpMemberOrFunctionOrValue as v -> not v.IsMutable
+            | _ -> true
 
-        let checkNotMutable (ident:Ident) =
-            async {
-                let! symbol =
-                    checkInfo.GetSymbolUseAtLocation(
-                        ident.idRange.StartLine, ident.idRange.EndColumn, "", [ident.idText])
+        let checkNotMutable (ident:Ident) = async {
+            let! symbol =
+                checkInfo.GetSymbolUseAtLocation(
+                    ident.idRange.StartLine, ident.idRange.EndColumn, "", [ident.idText])
 
-                let isNotMutable (symbol:FSharpSymbolUse) =
-                    match symbol.Symbol with
-                    | :? FSharpMemberOrFunctionOrValue as v -> not v.IsMutable
-                    | _ -> true
-
-                return
-                    match symbol with
-                    | Some(symbol) -> isNotMutable symbol
-                    | None -> false }
+            match symbol with
+            | Some(symbol) -> return isNotMutable symbol
+            | None -> return false
+        }
 
         let rec matchingIdentifier (bindingIdent:Ident) = function
             | SynExpr.Paren(expr, _, _, _) ->
@@ -44,14 +43,14 @@ let private checkForUselessBinding (checkInfo:FSharpCheckFileResults option) pat
             { Range = range
               Message = Resources.GetString("RulesUselessBindingError")
               SuggestedFix = None
-              TypeChecks = [checkNotMutable ident] })
+              TypeChecks = [ checkNotMutable ident ] })
         |> Option.toArray
     | _ -> Array.empty
 
 let private runner (args:AstNodeRuleParams) =
     match args.AstNode with
     | AstNode.Binding(SynBinding.Binding(_, _, _, isMutable, _, _, _, pattern, _, expr, range, _))
-            when Helper.Binding.isLetBinding args.NodeIndex args.SyntaxArray args.SkipArray
+            when Helper.Binding.isLetBinding args.NodeIndex args.SyntaxArray
                  && not isMutable ->
         checkForUselessBinding args.CheckInfo pattern expr range
     | _ ->