+++ src/FSharpLint.Core/Rules/Conventions/Naming/PublicValuesNames.fs
@@ -38,7 +38,7 @@ let private getIdentifiers (args:AstNodeRuleParams) =
         if not (isLiteral attributes) then
             match identifierTypeFromValData valData with
             | Value | Function ->
-                let isPublic = isPublic args.SyntaxArray args.SkipArray args.NodeIndex
+                let isPublic = isPublic args.SyntaxArray args.NodeIndex
                 getPatternIdents isPublic (getValueOrFunctionIdents args.CheckInfo) true pattern
             | _ -> Array.empty
         else