+++ src/FSharpLint.Core/Rules/Hints/HintsHelper.fs
@@ -21,16 +21,17 @@ open MergeSyntaxTrees
 /// Confirms if two parts of the ast look alike.
 /// This is required as hints can bind variables: the bound location needs to be compared to
 /// parts of the ast that the hint covers with the same variable.
-let private isMatch i j (nodeArray:AbstractSyntaxArray.Node []) (skipArray:AbstractSyntaxArray.Skip []) = 
-    let skipI = skipArray.[i].NumberOfChildren
-    let skipJ = skipArray.[j].NumberOfChildren
+let private isMatch i j (nodeArray:AbstractSyntaxArray.Node []) = 
+    let numChildrenI = nodeArray.[i].NumberOfChildren
+    let numChildrenJ = nodeArray.[j].NumberOfChildren
 
-    if skipI = skipJ then
-        Array.zip [|i..i + skipI|] [|j..j + skipJ|]
-        |> Array.forall (fun (i, j) -> 
-            i < nodeArray.Length && 
-            j < nodeArray.Length && 
-            nodeArray.[i].Hashcode = nodeArray.[j].Hashcode)
+    if numChildrenI = numChildrenJ then
+        let numChildren = numChildrenI
+
+        { 0..numChildren } |> Seq.forall (fun child -> 
+            i + child < nodeArray.Length && 
+            j + child < nodeArray.Length && 
+            nodeArray.[i + child].Hashcode = nodeArray.[j + child].Hashcode)
     else false
 
 let inline private isParen (node:AbstractSyntaxArray.Node) =
@@ -39,7 +40,7 @@ let inline private isParen (node:AbstractSyntaxArray.Node) =
     | _ -> false
 
 /// Compares the hint trie against a given location in the abstract syntax array.
-let rec checkTrie i trie (nodeArray:AbstractSyntaxArray.Node []) (skipArray:AbstractSyntaxArray.Skip []) (boundVariables:Dictionary<_, _>) notify =
+let rec checkTrie i trie (nodeArray:AbstractSyntaxArray.Node []) (boundVariables:Dictionary<_, _>) notify =
     trie.MatchedHint |> List.iter notify
 
     if i < nodeArray.Length then
@@ -47,10 +48,10 @@ let rec checkTrie i trie (nodeArray:AbstractSyntaxArray.Node []) (skipArray:Abst
 
         if isParen node then
             // Skip the paren.
-            checkTrie (i + 1) trie nodeArray skipArray boundVariables notify
+            checkTrie (i + 1) trie nodeArray boundVariables notify
         else
             match trie.Edges.Lookup.TryGetValue node.Hashcode with
-            | true, trie -> checkTrie (i + 1) trie nodeArray skipArray boundVariables notify
+            | true, trie -> checkTrie (i + 1) trie nodeArray boundVariables notify
             | false, _ -> ()
 
         trie.Edges.AnyMatch
@@ -58,11 +59,11 @@ let rec checkTrie i trie (nodeArray:AbstractSyntaxArray.Node []) (skipArray:Abst
             match var with
             | Some(var) -> 
                 match boundVariables.TryGetValue var with
-                | true, varI when isMatch varI i nodeArray skipArray  -> 
-                    checkTrie (i + skipArray.[i].NumberOfChildren + 1) trie nodeArray skipArray boundVariables notify
+                | true, varI when isMatch varI i nodeArray  -> 
+                    checkTrie (i + node.NumberOfChildren + 1) trie nodeArray boundVariables notify
                 | false, _ -> 
                     boundVariables.Add(var, i)
-                    checkTrie (i + skipArray.[i].NumberOfChildren + 1) trie nodeArray skipArray boundVariables notify
+                    checkTrie (i + node.NumberOfChildren + 1) trie nodeArray boundVariables notify
                 | true, _ -> ()
-            | None -> checkTrie (i + skipArray.[i].NumberOfChildren + 1) trie nodeArray skipArray boundVariables notify)
+            | None -> checkTrie (i + node.NumberOfChildren + 1) trie nodeArray boundVariables notify)